Finish basic styling og the webpage
I finished the overall styling of the webpage.
I need to work on the maximum width of the page,
thought. On large screens the header section is
spaced apart too much. However, applying a maximum
width to the header section and aligning items
centrally collapses the header horizontally.
Need to work on this.